Be robust to invalid utf-8 characters in task db
Sometimes, there may exist non-printable characters in the taskwarrior `pending.data` file, for example due to an emoji added to the task description but is not yet properly parsed by python. In these cases, we'd want `tasklib` not to crash but rather ignore it and keep parsing the results of the command.